LayoutTests/crypto/subtle/wrapKey-check-usage-expected.txt

CONSOLE MESSAGE: line 1: Key usages do not include 'wrapKey'
Verify that usage is respected in wrapKey operation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing an RSA key...
Importing an AES key...
Wrapping it...
PASS crypto.subtle.wrapKey("raw", aesKey, publicKey, wrapAlgorithm) threw exception Error: NotSupportedError: DOM Exception 9.
PASS successfullyParsed is true

TEST COMPLETE
